chore: move the parameters type from type to api-verbs
