+++ apps/web/src/schema/identity/service-account/api-verbs/list.ts
@@ -1,0 +1,11 @@
+export interface ServiceAccountListParameter {
+    domain_id?: string;
+    query?: any;
+    service_account_id?: string;
+    name?: string;
+    provider?: string;
+    has_secret?: boolean;
+    workspace_id?: string;
+    project_id?: string;
+    trusted_service_account_id?: string;
+}
